(clean) Refactor geo query functions to be more modular.
