+++ server/workers/openDataCaller.js
@@ -1,7 +1,6 @@
 var request = require('request');
 var db = require('../models/crime');
 var dbBox = require('../models/boxCrime');
-console.log('in open caller')
 
 var requestQuery =
 'https://data.sfgov.org/resource/9v2m-8wqu.json?\
@@ -19,13 +18,11 @@ request(requestQuery, function(err, res, body) {
     var results = JSON.parse(body);
     db.clearDatabase(function(err) {
       db.storeOpenData(results, function(err) {
-        console.log('stored data normally')
         if (err) {
           console.error(err);
         } else {
           dbBox.clearBoxDatabase(function(err) {
             dbBox.storeOpenDataInBoxes(results, function(err) {
-              console.log('stored data in boxes')
               if (err) {
                 console.error(err);
               }
